323 done

+++ src/com/bdg/nhakobyan/lesson_6/exercise_230.java
@@ -4,7 +4,7 @@ public class exercise_230 {
     public static void main(String[] args) {
         double [] arr = {10.5 ,35.3, 7.8,0,1,12,98,4,8,4,0.6,9};
         int k = 8;
-        int amboxch = 0;
+        double amboxch = 0;
         int sum = 0;
         int counter = 0;
         for (int i = 0; i < arr.length ; i++) {
